requirements.txt
# Core dependencies
neo4j==5.13.0
py2neo==2021.2.4
pyarrow<12.0.0,>=3.0.0
pandas==2.1.1
pydantic==2.5.2
pydantic-settings==2.0.3

# API and Web
fastapi==0.104.1
uvicorn==0.23.2
httpx==0.25.0

# OCSF Schema
jsonschema==4.19.1
pyyaml==6.0.1

# Data Processing
apache-beam==2.50.0
confluent-kafka==2.3.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
docker==7.0.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Development dependencies
black==23.10.1
isort==5.12.0
flake8==6.1.0 